+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,58 +1,19 @@
 package roomescape.controller;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationRepository;
-import roomescape.dto.ReservationDto;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/admin")
 public class AdminController {
-    private final ReservationRepository reservationRepository;
-
-    @Autowired
-    public AdminController(ReservationRepository reservationRepository) {
-        this.reservationRepository = reservationRepository;
-    }
-
-    @GetMapping("/admin")
+    @GetMapping
     public String admin() {
         return "admin/index";
     }
 
-    @GetMapping("/admin/reservation")
+    @GetMapping("/reservation")
     public String reservation() {
         return "admin/reservation-legacy";
-    }
-
-    @GetMapping("/reservations")
-    @ResponseBody
-    public ResponseEntity<List<Reservation>> reservations() {
-        List<Reservation> reservations = reservationRepository.findAll();
-        return ResponseEntity.ok(reservations);
-    }
-
-    @PostMapping("/reservations")
-    @ResponseBody
-    public ResponseEntity<Reservation> createReservation(@RequestBody ReservationDto reservationDto) {
-        Reservation newReservation = reservationRepository.save(
-                new Reservation(reservationDto.name(), reservationDto.date(), reservationDto.time()));
-        return ResponseEntity.ok(newReservation);
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable(value = "id") Long id) {
-        reservationRepository.deleteById(id);
-        return ResponseEntity.ok().build();
     }
 }
