+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,0 +1,44 @@
+package roomescape.controller;
+
+import java.util.List;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationService;
+import roomescape.dto.ReservationDto;
+
+@Controller
+@RequestMapping("/reservations")
+public class ReservationController {
+    private final ReservationService reservationService;
+
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Reservation>> reservations() {
+        List<Reservation> reservations = reservationService.findReservations();
+        return ResponseEntity.ok(reservations);
+    }
+
+    @PostMapping
+    public ResponseEntity<Reservation> createReservation(@RequestBody ReservationDto reservationDto) {
+        Reservation newReservation = reservationService.create(reservationDto);
+        return ResponseEntity.ok(newReservation);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable(value = "id") Long id) {
+        reservationService.cancel(id);
+        return ResponseEntity.ok().build();
+    }
+}
